fix(c-core): use long long instead of long to avoid conversion issues on Windows
- Changed variable types from long to long long to ensure proper handling of large values on Windows.
- This prevents potential issues where long is 32 bits on Windows, which could lead to incorrect conversions and overflow.

+++ ext/fast_xirr/common.h
@@ -8,8 +8,8 @@ typedef struct {
     time_t date;
 } CashFlow;
 
-double npv(double rate, CashFlow *cashflows, long count, time_t min_date);
+double npv(double rate, CashFlow *cashflows, long long count, time_t min_date);
 
-int find_bracketing_interval(CashFlow *cashflows, long count, double *low, double *high);
+int find_bracketing_interval(CashFlow *cashflows, long long count, double *low, double *high);
 
 #endif
